requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
backports.asyncio.runner==1.2.0
bcrypt==4.0.1
boto3==1.26.73
botocore==1.29.73
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==7.1.2
cloudinary==1.44.1
cryptography==41.0.4
dnspython==2.7.0
docopt==0.6.2
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.8
Flask==1.1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.4
iniconfig==2.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==1.0.1
MarkupSafe==2.0.1
numpy==2.0.2
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pipreqs==0.4.11
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypi-xmlrpc==2020.12.3
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.5.0
pytz==2025.2
requests==2.28.2
rsa==4.9.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.46
starlette==0.49.3
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.14
uvicorn==0.39.0
Werkzeug==1.0.1
yarg==0.1.9
